requirements.txt
fastapi==0.68.0
sqlalchemy==1.4.46
pillow==10.3.0
ezdxf==1.3.0
trimesh==4.2.0
pandas==2.0.3
python-dotenv==1.0.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
numpy==1.26.4
setuptools==68.2.2
cython==3.0.11
